refactor into data source logic
1. moved google worksheet into its own file
2. make a data source object that manages the three
datasets: working, history, and current
3.  performed type conversion on all of them.

It's important to hide the load behind a property so we can
cache the results.  The working checks are going to need
the history (or at least data derived from the history).

Also, we may want to cache the results and only rerun on change.