restructured to use a 'logger' model
Most simple check routines return either a warning or an error
but the minute you look over time, you need to be able to return
multiple errors.

So I switched to a logger model where we pass in a ResultLog
object and write to it instead of returning the errors from the
check routines.