added an increasing value check
It checks to see if values changed against the prior data.  If they
are the same, something is probably wrong.  It also checks for
an expected increase.

Right now, it has hard-coded thresholds that should be estimated
from history.

It also doesn't compare date well enough.  It assumes the working
sheet is always one day ahead of the published results.  This is
not true from 4PM to maybe 10AM.  We need to compute if
the worksheet is published and use an offset=1 to get the right
data to compare against.